+++ examples/tasks/main.cpp
@@ -52,7 +52,8 @@ int main(int argc, char* argv[]) {
 
     // create_obx_model() provided by objectbox-model.h
     // obx interface contents provided by objectbox.hpp
-    obx::Options options(create_obx_model());
+    obx::Options options;
+    options.model(create_obx_model());
 
     if (int err = processArgs(argc, argv, options)) {
         return err;